# schemas/openapi/odyssey-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "DiagnosticsResponse": {
        "properties": {
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Capabilities",
            "type": "object"
          },
          "configuration": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Configuration",
            "type": "object"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "telemetry": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Telemetry",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "service",
          "version",
          "configuration",
          "capabilities",
          "telemetry"
        ],
        "title": "DiagnosticsResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SyncConflictListResponse": {
        "additionalProperties": false,
        "properties": {
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/SyncConflictDetail"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "pending_count": {
            "minimum": 0.0,
            "title": "Pending Count",
            "type": "integer"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "conflicts",
          "pending_count",
          "server_time"
        ],
        "title": "SyncConflictListResponse",
        "type": "object"
      },
      "SyncConflictResolutionRequest": {
        "additionalProperties": false,
        "properties": {
          "client_schema_version": {
            "minimum": 1.0,
            "title": "Client Schema Version",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "device_sequence": {
            "minimum": 1.0,
            "title": "Device Sequence",
            "type": "integer"
          },
          "expected_current_revision": {
            "minimum": 1.0,
            "title": "Expected Current Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "merged_document": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged Document"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "sensitivity_class": {
            "$ref": "#/components/schemas/DataClass",
            "default": "private"
          },
          "strategy": {
            "$ref": "#/components/schemas/ConflictResolutionStrategy"
          }
        },
        "required": [
          "operation_id",
          "device_id",
          "device_sequence",
          "client_schema_version",
          "expected_current_revision",
          "strategy",
          "created_at"
        ],
        "title": "SyncConflictResolutionRequest",
        "type": "object"
      },
      "SyncConflictResolutionResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted_operation": {
            "$ref": "#/components/schemas/AcceptedOperation"
          },
          "conflict_id": {
            "format": "uuid",
            "title": "Conflict Id",
            "type": "string"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string"
          },
          "resolution_id": {
            "format": "uuid",
            "title": "Resolution Id",
            "type": "string"
          },
          "server_schema_version": {
            "minimum": 1.0,
            "title": "Server Schema Version",
            "type": "integer"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "strategy": {
            "$ref": "#/components/schemas/ConflictResolutionStrategy"
          }
        },
        "required": [
          "resolution_id",
          "conflict_id",
          "status",
          "strategy",
          "accepted_operation",
          "next_cursor",
          "server_time",
          "server_schema_version"
        ],
        "title": "SyncConflictResolutionResponse",
        "type": "object"
      },
      "SyncConflictStatusFilter": {
        "enum": [
          "pending",
          "resolved",
          "all"
        ],
        "title": "SyncConflictStatusFilter",
        "type": "string"
      },
...
      "SyncDeviceDiagnostics": {
        "additionalProperties": false,
        "properties": {
          "attachment_backlog": {
            "minimum": 0.0,
            "title": "Attachment Backlog",
            "type": "integer"
          },
          "client_schema_version": {
            "minimum": 1.0,
            "title": "Client Schema Version",
            "type": "integer"
          },
          "clock_skew_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clock Skew Seconds"
          },
          "device_cursor": {
            "title": "Device Cursor",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "diagnostics_reported_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostics Reported At"
          },
          "diagnostics_stale": {
            "title": "Diagnostics Stale",
            "type": "boolean"
          },
          "last_device_sequence": {
            "minimum": 0.0,
            "title": "Last Device Sequence",
            "type": "integer"
          },
          "last_successful_pull_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Successful Pull At"
          },
          "last_successful_push_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Successful Push At"
          },
          "oldest_unsynced_operation_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Unsynced Operation At"
          },
          "operations_queued": {
            "minimum": 0.0,
            "title": "Operations Queued",
            "type": "integer"
          },
          "schema_compatibility": {
            "$ref": "#/components/schemas/SchemaCompatibility"
          },
          "server_cursor": {
            "title": "Server Cursor",
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "client_schema_version",
          "schema_compatibility",
          "last_successful_push_at",
          "last_successful_pull_at",
          "operations_queued",
          "oldest_unsynced_operation_at",
          "attachment_backlog",
          "last_device_sequence",
          "device_cursor",
          "server_cursor",
          "clock_skew_seconds",
          "diagnostics_reported_at",
          "diagnostics_stale"
        ],
        "title": "SyncDeviceDiagnostics",
        "type": "object"
      },
      "SyncDeviceDiagnosticsInput": {
        "additionalProperties": false,
        "properties": {
          "attachment_backlog": {
            "minimum": 0.0,
            "title": "Attachment Backlog",
            "type": "integer"
          },
          "client_schema_version": {
            "minimum": 1.0,
            "title": "Client Schema Version",
            "type": "integer"
          },
          "device_cursor": {
            "title": "Device Cursor",
            "type": "string"
          },
          "oldest_unsynced_operation_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Unsynced Operation At"
          },
          "operations_queued": {
            "minimum": 0.0,
            "title": "Operations Queued",
            "type": "integer"
          }
        },
        "required": [
          "client_schema_version",
          "device_cursor",
          "operations_queued",
          "attachment_backlog"
        ],
        "title": "SyncDeviceDiagnosticsInput",
        "type": "object"
      },
      "SyncDiagnosticsResponse": {
        "additionalProperties": false,
        "properties": {
          "devices": {
            "items": {
              "$ref": "#/components/schemas/SyncDeviceDiagnostics"
            },
            "title": "Devices",
            "type": "array"
          },
          "minimum_client_schema_version": {
            "minimum": 1.0,
            "title": "Minimum Client Schema Version",
            "type": "integer"
          },
          "pending_attachment_uploads": {
            "minimum": 0.0,
            "title": "Pending Attachment Uploads",
            "type": "integer"
          },
          "pending_conflicts": {
            "minimum": 0.0,
            "title": "Pending Conflicts",
            "type": "integer"
          },
          "pending_outbox_jobs": {
            "minimum": 0.0,
            "title": "Pending Outbox Jobs",
            "type": "integer"
          },
          "repair": {
            "$ref": "#/components/schemas/SyncRepairOptions"
          },
          "server_cursor": {
            "title": "Server Cursor",
            "type": "string"
          },
          "server_schema_version": {
            "minimum": 1.0,
            "title": "Server Schema Version",
            "type": "integer"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "sync_pull_enabled": {
            "title": "Sync Pull Enabled",
            "type": "boolean"
          },
          "sync_push_enabled": {
            "title": "Sync Push Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "server_time",
          "server_cursor",
          "server_schema_version",
          "minimum_client_schema_version",
          "pending_conflicts",
          "pending_attachment_uploads",
          "pending_outbox_jobs",
          "sync_push_enabled",
          "sync_pull_enabled",
          "devices",
          "repair"
        ],
        "title": "SyncDiagnosticsResponse",
        "type": "object"
      },
...
      "SyncPullResponse": {
        "additionalProperties": false,
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/SyncChange"
            },
            "title": "Changes",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "minimum_client_schema_version": {
            "minimum": 1.0,
            "title": "Minimum Client Schema Version",
            "type": "integer"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "string"
          },
          "server_schema_version": {
            "minimum": 1.0,
            "title": "Server Schema Version",
            "type": "integer"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          }
        },
        "required": [
          "changes",
          "next_cursor",
          "has_more",
          "server_time",
          "server_schema_version",
          "minimum_client_schema_version"
        ],
        "title": "SyncPullResponse",
        "type": "object"
      },
      "SyncPushRequest": {
        "additionalProperties": false,
        "properties": {
          "base_cursor": {
            "title": "Base Cursor",
            "type": "string"
          },
          "client_schema_version": {
            "minimum": 1.0,
            "title": "Client Schema Version",
            "type": "integer"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/SyncOperationInput"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Operations",
            "type": "array"
          }
        },
        "required": [
          "device_id",
          "client_schema_version",
          "base_cursor",
          "operations"
        ],
        "title": "SyncPushRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Private local-first personal navigation API",
    "title": "Odyssey API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/diagnostics": {
      "get": {
        "operationId": "diagnostics_v1_admin_diagnostics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostics",
        "tags": [
          "system"
        ]
      }
    },
...
    "/v1/sync/changes": {
      "get": {
        "operationId": "changes_v1_sync_changes_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": "c_0",
              "minLength": 3,
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Odyssey-Device-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Odyssey-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Changes",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/conflicts": {
      "get": {
        "operationId": "conflicts_v1_sync_conflicts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SyncConflictStatusFilter",
              "default": "pending"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflictListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Conflicts",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/conflicts/{conflict_id}/resolve": {
      "post": {
        "operationId": "resolve_conflict_v1_sync_conflicts__conflict_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "conflict_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conflict Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncConflictResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncConflictResolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Conflict",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/devices/{device_id}/diagnostics": {
      "put": {
        "operationId": "report_device_diagnostics_v1_sync_devices__device_id__diagnostics_put",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncDeviceDiagnosticsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncDeviceDiagnostics"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Report Device Diagnostics",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/diagnostics": {
      "get": {
        "operationId": "diagnostics_v1_sync_diagnostics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncDiagnosticsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Diagnostics",
        "tags": [
          "sync"
        ]
      }
    },
    "/v1/sync/push": {
      "post": {
        "operationId": "push_v1_sync_push_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncPushRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPushResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Push",
        "tags": [
          "sync"
        ]
      }
    }
  }
}